requirements.txt
py-cui~=0.1.4
qiskit~=0.34.2
numpy~=1.22.0
antlr4-python3-runtime~=4.12
setuptools~=52.0.0
